fix(review): move recipe failure rendering to server
